Make sure input argument "devid" to iio_open is bounded
The "devid" input argument to iio_open() is used in iio_read_hostinfo()
to compose the filename string, which is a fixed sized char array.  This
means that the client caller is trusted to not pass a string that will
result in an out-of-bound memory write.

Validate the input size (I arbitrarily chose the size of 256, determined
conservatively from the client code in the proposed QEMU driver that
caps the size to 37 bytes).  This will now exit with error in the
iio_open() if devid size exceeds the limits, and in the
iio_read_hostinfo() call the string copy is made safe to make sure the
composed string never exceeds the array size.