test server: fix double-free on file pointer.
Both the read and write functions are using a static global file
pointer.  This means that the pointer gets clobbered during
multi-threaded operation.

Since the referenced file pointer is opened, operated on, and closed all
in the same function call, there is no reason for it to not be scoped to
the function.

(This was noticed because when running qemu with a raw image on the
 test-server, the test-server would abort() on a double-free).